added card component to it's own folder
