+++ src/App.js
@@ -1,5 +1,6 @@
 import './App.css';
 import React, { useState } from "react";
+import Card from "../src/components/cards/Card.js"
 
 function App() {
   const cardContent = {
@@ -21,23 +22,6 @@ function App() {
   }
 
   const shuffledCards = shuffle(Object.entries(cardContent))
-
-  function Card({cardName, cardText}) {
-
-    const [flip, setFlip] = useState(false);
-
-    return (
-      <div className={`card ${flip ? 'flip' : ''}`}>
-        <div className="front" onClick={() => setFlip(!flip)}>
-          <h3>click me</h3>
-        </div>
-        <div className="back" onClick={() => setFlip(!flip)}>
-          <h3>{cardName}</h3>
-          <p>{cardText}</p>
-        </div>
-      </div>
-    );
-  }
 
   return (
     <div className="cards_container">
